Add files via upload
this is a basic chatbot designed with the help of python, nltk libraries. It gives the user tips on the basis of how they are feeling. For eg- if the user is feeling sad then the chatbot will give the tips for what all the user can do in order to feel better.